requirements.txt
langchain>=0.2.0
langchain-openai
langchain-groq
langchain-community
langchain-experimental
python-dotenv
openai
groq
tavily-python
pydantic>=2.0
langgraph
# langgraph_core
pydantic>=2.0
pydantic-settings
FastAPI
uvicorn
streamlit
requests
langchain_huggingface
huggingface_hub